Make tests a JS file
